requirements.txt
Django==3.2
gunicorn==20.1.0
python-dotenv==0.20.0
mssql-django==1.1.3
pyodbc==4.0.34
whitenoise==6.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0 
pyjwt==2.4.0
django-crispy-forms==1.14.0




